# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/add": {
            "post": {
                "description": "Добавляет новую задачу ToDo.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Post"
                ],
                "summary": "Добавление новой задачи",
                "parameters": [
                    {
                        "description": "Новая задача",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ToDoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ToDo"
                        }
                    }
                }
            }
        },
        "/change/{title}": {
            "patch": {
                "description": "Изменяет задачу ToDo по заголовку.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Patch"
                ],
                "summary": "Изменение задачи по заголовку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Заголовок задачи",
                        "name": "title",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Обновленная задача",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ToDoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ToDo"
                        }
                    }
                }
            }
        },
        "/date/{flag}": {
            "get": {
                "description": "Получает список задач ToDo с фильтром по флагу выполнения и сортировкой по дате.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get"
                ],
                "summary": "Отображение списка задач с фильтром по флагу выполнения и сортировкой по дате",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Флаг выполнения задачи (true/false)",
                        "name": "flag",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ToDo"
                            }
                        }
                    }
                }
            }
        },
        "/del/{title}": {
            "delete": {
                "description": "Удаляет задачу ToDo по заголовку.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Delete"
                ],
                "summary": "Удаление задачи по заголовку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Заголовок задачи",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "todo deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/false": {
            "get": {
                "description": "Получает список задач ToDo с фильтром по флагу выполнения и пагинацией.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get"
                ],
                "summary": "Отображение списка задач с фильтром по флагу выполнения и пагинацией",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Флаг выполнения задачи (true/false)",
                        "name": "flag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Номер страницы для пагинации",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ToDo"
                            }
                        }
                    }
                }
            }
        },
        "/flag/{title}": {
            "get": {
                "description": "Меняет флаг выполнения задачи ToDo по ее заголовку",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get"
                ],
                "summary": "Изменение флага выполнения задачи ToDo",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Заголовок задачи",
                        "name": "title",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.ToDo"
                        }
                    }
                }
            }
        },
        "/show": {
            "get": {
                "description": "Show all.",
                "tags": [
                    "Get"
                ],
                "summary": "Show all ToDo's",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ToDo"
                            }
                        }
                    }
                }
            }
        },
        "/true": {
            "get": {
                "description": "Получает список задач ToDo с фильтром по флагу выполнения и пагинацией.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Get"
                ],
                "summary": "Отображение списка задач с фильтром по флагу выполнения и пагинацией",
                "parameters": [
                    {
                        "type": "boolean",
                        "description": "Флаг выполнения задачи (true/false)",
                        "name": "flag",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Номер страницы для пагинации",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ToDo"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.ToDo": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "flag": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ToDoRequest": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "desc": {
                    "type": "string"
                },
                "flag": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}